res reorganized the endian conversion stuff a bit; the routines are   now wrapped into classes. Also, they are now named to reflect that   specifically sized types are converted. Last but not least,   encourage use of IEEE floats instead of a custom floating-point   representation.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22468 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/loader/bmp/bmpimage.cpp
@@ -224,7 +224,7 @@ csPtr<iDataBuffer> csBMPImageIO::Save (iImage *Image, iImageIO::FileFormatDescri
   bmpHeader hdr;
   hdr.bfTypeLo = 'B';
   hdr.bfTypeHi = 'M';
-  hdr.bfSize = (uint32)csLittleEndianLong (len);
+  hdr.bfSize = (uint32)csLittleEndianLong ((uint32)len);
   hdr.bfRes1 = 0;
   hdr.bfRes2 = 0;
   hdr.bfOffBits = csLittleEndianLong (sizeof (bmpHeader)-2 + 256*(palette?4:0));
